update getImages() call to accept a callback that fires only once
Before, it was `getImages(urls, options, callback);` now it's `getImages(urls, options, callback, singleCallback);`. The `callback` gets called only once after all images have loaded; the `singleCallback` gets fired on each image load, similar to the previous behavior. The `callback` gets called with an argument that is an array of all of the instantiated `Image` objects.