Module is es2015-compatible.
The current module entry point is the source code itself.

Rollup has no problem using this, but for Webpack (and other bundling
tools that were not designed for es modules first), the exported module
should be at least minimally transpiled to be compliant with these
tools (currently Webpack will not bundle this file).

https://github.com/facebook/create-react-app/blob/master/packages/react-scripts/template/README.md#npm-run-build-fails-to-minify